+++ apps/backend/src/libs/modules/server-application/libs/types/types.ts
@@ -1,4 +1,6 @@
 export { type APIDoc } from "./api-document.type.js";
+export { type ServerApplicationApiConstructorParams } from "./server-application-api-constructor-parameters.type.js";
 export { type ServerApplicationApi } from "./server-application-api.type.js";
 export { type ServerApplicationRouteParameters } from "./server-application-route-parameters.type.js";
 export { type ServerApplication } from "./server-application.type.js";
+export { type WhiteRoute } from "~/libs/modules/plugins/auth/auth.plugin.js";