+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -1,20 +1,36 @@
+import { APIPath, AuthApiPath } from "~/libs/enums/enums.js";
 import { config } from "~/libs/modules/config/config.js";
 import { database } from "~/libs/modules/database/database.js";
 import { logger } from "~/libs/modules/logger/logger.js";
 import { authController } from "~/modules/auth/auth.js";
 import { pointsOfInterestController } from "~/modules/points-of-interest/points-of-interest.js";
+import { routeCategoryController } from "~/modules/route-categories/route-categories.js";
 import { userController } from "~/modules/users/users.js";
 
 import { BaseServerApplicationApi } from "./base-server-application-api.js";
 import { BaseServerApplication } from "./base-server-application.js";
 
 const apiV1 = new BaseServerApplicationApi(
-	"v1",
-	config,
+	{
+		config,
+		version: "v1",
+		whiteRoutes: [
+			{
+				method: "POST",
+				path: `${APIPath.AUTH}${AuthApiPath.SIGN_UP}`,
+			},
+			{
+				method: "POST",
+				path: `${APIPath.AUTH}${AuthApiPath.SIGN_IN}`,
+			},
+		],
+	},
 	...authController.routes,
+	...routeCategoryController.routes,
 	...userController.routes,
 	...pointsOfInterestController.routes,
 );
+
 const serverApplication = new BaseServerApplication({
 	apis: [apiV1],
 	config,