+++ apps/backend/src/modules/users/users.ts
@@ -11,9 +11,14 @@ const userController = new UserController(logger, userService);
 
 export { userController, userService };
 export {
-	type UserGetAllItemResponseDto,
+	type UserGetByIdItemResponseDto,
+	type UserSignInRequestDto,
+	type UserSignInResponseDto,
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
 } from "./libs/types/types.js";
-export { userSignUpValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
+export {
+	userSignInValidationSchema,
+	userSignUpValidationSchema,
+} from "./libs/validation-schemas/validation-schemas.js";
 export { type UserService } from "./user.service.js";