+++ apps/frontend/src/libs/components/carousel/libs/types/carousel-reference.type.ts
@@ -1,0 +1,11 @@
+type CarouselReference = {
+	element: React.RefObject<HTMLDivElement | null>;
+	isAnimating: React.RefObject<boolean>;
+	isDragging: React.RefObject<boolean>;
+	momentumID: React.RefObject<null | number>;
+	scrollStart: React.RefObject<number>;
+	startX: React.RefObject<number>;
+	velocity: React.RefObject<number>;
+};
+
+export { type CarouselReference };