+++ apps/frontend/src/pages/auth/auth.tsx
@@ -1,6 +1,7 @@
 import { Navigate } from "react-router";
 
 import logo from "~/assets/images/logo.svg";
+import { Link } from "~/libs/components/components.js";
 import { AppRoute } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
@@ -58,14 +59,16 @@ const Auth = (): React.JSX.Element => {
 	);
 
 	if (hasUser) {
-		return <Navigate to={AppRoute.ROOT} />;
+		return <Navigate to={AppRoute.APP} />;
 	}
 
 	return (
 		<main className={styles["container"]}>
 			<div className={styles["left-panel"]}>
 				<div className={styles["logo"]}>
-					<img alt="SmartScapes" height="24" src={logo} width="136" />
+					<Link to={AppRoute.ROOT}>
+						<img alt="SmartScapes" height="24" src={logo} width="136" />
+					</Link>
 				</div>
 
 				{handleFormRender(pathname)}