+++ packages/shared/src/modules/users/users.ts
@@ -1,11 +1,13 @@
 export { UserExceptionMessage, UsersApiPath } from "./libs/enums/enums.js";
 export {
 	type UserAuthResponseDto,
-	type UserGetAllItemResponseDto,
+	type UserGetByIdItemResponseDto,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
 } from "./libs/types/types.js";
-export { userSignUp as userSignUpValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
-export { userSignIn as userSignInValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
+export {
+	userSignIn as userSignInValidationSchema,
+	userSignUp as userSignUpValidationSchema,
+} from "./libs/validation-schemas/validation-schemas.js";